@@ -14,7 +14,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3.0",
+        "php": ">=7.4.0",
         "myclabs/deep-copy": "^1.10",
         "phpunit/phpunit": "^8.0|^9.0"
     },
